requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.12.14
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.1
greenlet==3.1.1
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
Mako==1.3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mysqlclient==2.2.6
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
